Fixed unit tests and renamed a forgotten class to be compliant with code standard
