+++ Sources.cmake
@@ -29,11 +29,11 @@ set (sources
     src/libc-stub.cpp
     src/main.cpp
     src/camera.cpp
-    src/vector-2d.cpp
+    src/vector_2d.cpp
     src/mock.cpp
-    src/Image.cpp
-    src/ColorPixel.cpp
-    src/ColorImage.cpp
-    src/GreyscalePixel.cpp
-    src/GreyscaleImage.cpp
+    src/image.cpp
+    src/color_pixel.cpp
+    src/color_image.cpp
+    src/greyscale_pixel.cpp
+    src/greyscale_image.cpp
 )
