+++ test/test_main.cpp
@@ -1,7 +1,7 @@
-z#define CATCH_CONFIG_MAIN // This tells Catch to provide a main() - only do this in one cpp file
+#define CATCH_CONFIG_MAIN // This tells Catch to provide a main() - only do this in one cpp file
 #include "../src/vector_2d.cpp"
 
-#include "../src/Image.cpp"
+#include "../src/image.cpp"
 
 #include "../src/mock.cpp"
